Rename notebook to unrevised
I realized that I really don't need everything I stuffed into my
notebook and that I'm stressing myself over nothing thinking that I have
to go through and clean up the entire thing. Instead, this is my
unrevised notebook, it's messy but hard work, lots of stuff.
I'm going to copy this notebook and make a revised version that's
cleaner and only include the details that matters. Then after that I'm
going to make a model notebook that's just for running the model.